+++ .changeset/strange-boxes-fry.md
@@ -1,0 +1,3 @@
+---
+'@finos/legend-extension-dsl-data-space': patch
+---
